Add run_to_report to mvpa postproc
Remove code from if statement for across subjects in batch script, so
that it can work with any analysis. Make separate function, and call it
in single subject script, which should restructure mvpa results so that it only contains the variables of interest.
Update the saved name so that it reflects this information.