Add code to train and test mvpa on different TRs
If the training and testing TRs are specified as different, assert that
all selectors are in {0, 1, 2}, and create temporally condensed data for
the training set and the testing set. Use horzcat to append testing data
onto training data. If the traintest specification sets trials to 0,
don't keep them in the patterns, regressors, or selectors.