Condense patterns based train test in mvpa
Use which_traintest to further condense temporally condensed data, as
opposed to simply taking out the rest timepoints in the temporally
condensed data. This will make the patterns, regressors, and selectors
indices match and will avoid the problem of adding regressors during
outlier detection. Previously, outlier detection would find patterns not
referenced by the regressors and then try to set them to 0 in the
regressors, which required the regressors to be extended by many
indices.